+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is roughly based on [Keep a Changelog](https://keepachangelog.com/en/
 
 - Bump minimum required Python version to 3.11 ([#127](https://github.com/Open-EO/openeo-aggregator/issues/127), [#174](https://github.com/Open-EO/openeo-aggregator/issues/174))
 - Add title, description, online/offline status and last_status_check to federation listing in capabilities document ([#22](https://github.com/Open-EO/openeo-aggregator/issues/22))
-
+- Provide "federation:missing" on `/file_formats` ([#124](https://github.com/Open-EO/openeo-aggregator/issues/124))
 
 ## 0.43.0